# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JernihAI API
  description: 'API backend JernihAI — platform peningkatan kualitas gambar.


    Dokumentasi API Publik B2B (FR-14) untuk developer: `docs/API_B2B.md`. Spesifikasi
    OpenAPI lengkap: `docs/api/openapi.yaml` (regenerate via `python api/scripts/export_openapi.py`).'
  version: 0.1.0
paths:
...
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Daftar akun baru
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login email + password
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/auth/google:
    get:
      tags:
      - auth
      summary: Mulai login Google (redirect ke Google)
      operationId: google_login_api_v1_auth_google_get
      responses:
        '302':
          description: Redirect ke Google OAuth consent
          content:
            application/json:
              schema: {}
        '503':
          description: Google login belum dikonfigurasi (GOOGLE_CLIENT_ID kosong)
  /api/v1/auth/google/callback:
    get:
      tags:
      - auth
      summary: Callback Google — tukar code dengan token lalu set cookie
      operationId: google_callback_api_v1_auth_google_callback_get
      parameters:
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: error
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '303':
          description: Redirect ke web setelah login sukses
        '400':
          description: State OAuth tidak valid / email tidak tersedia
        '401':
          description: Gagal menukar kode Google
        '403':
          description: Akun dinonaktifkan
        '503':
          description: Google login belum dikonfigurasi
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
...
    RegisterRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        privacy_consent:
          type: boolean
          title: Privacy Consent
          description: Persetujuan kebijakan privasi (FR-07)
          default: false
      type: object
      required:
      - email
      - password
      - name
      title: RegisterRequest
...
    UserOut:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          format: email
          title: Email
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        provider:
          type: string
          title: Provider
        privacy_consent_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Privacy Consent At
        is_admin:
          type: boolean
          title: Is Admin
          default: false
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - email
      - name
      - provider
      - created_at
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...